@@ -1,19 +1,17 @@
 buildscript {
-    ext.kotlin_version = '1.3.31'
-    ext.support_version = '28.0.0'
-    ext.retrofit_version = '2.5.0'
-    ext.dagger_version = '2.22.1'
-    ext.okhttp3_version = '3.12.0'
-    ext.mockito_version = '2.27.0'
-    ext.arch_components_version = '1.1.1'
-    ext.leakcanary_version = '1.6.3'
+    ext.kotlin_version = '1.4.31'
+    ext.retrofit_version = '2.9.0'
+    ext.dagger_version = '2.33'
+    ext.okhttp3_version = '4.9.1'
+    ext.mockito_version = '3.8.0'
+    ext.leakcanary_version = '2.6'
 
     repositories {
         google()
         jcenter()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.4.1'
+        classpath 'com.android.tools.build:gradle:4.1.2'
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
         classpath 'com.cookpad.android.licensetools:license-tools-plugin:1.7.0'
         classpath "org.jetbrains.kotlin:kotlin-allopen:$kotlin_version"
