@@ -1,12 +1,12 @@
 package de.metzgore.beansplan.shared;
 
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.RecyclerView;
 
 import de.metzgore.beansplan.databinding.ListItemShowBinding;
 
 public class ShowViewHolder extends RecyclerView.ViewHolder {
 
-    private ListItemShowBinding binding;
+    private final ListItemShowBinding binding;
 
     public ShowViewHolder(ListItemShowBinding binding) {
         super(binding.getRoot());
