@@ -1,7 +1,7 @@
 package de.metzgore.beansplan.util;
 
 import android.content.Context;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 import android.text.format.DateUtils;
 
 import java.util.Date;
