@@ -1,9 +1,9 @@
 package de.metzgore.beansplan.util
 
-import android.arch.lifecycle.LiveData
-import android.arch.lifecycle.MediatorLiveData
-import android.support.annotation.MainThread
-import android.support.annotation.WorkerThread
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MediatorLiveData
+import androidx.annotation.MainThread
+import androidx.annotation.WorkerThread
 import de.metzgore.beansplan.AppExecutors
 import de.metzgore.beansplan.api.ApiEmptyResponse
 import de.metzgore.beansplan.api.ApiErrorResponse
