@@ -3,11 +3,9 @@ package de.metzgore.beansplan
 import android.app.Activity
 import android.app.Application
 import android.content.BroadcastReceiver
-import com.squareup.leakcanary.LeakCanary
 import dagger.android.AndroidInjector
 import dagger.android.DispatchingAndroidInjector
-import dagger.android.HasActivityInjector
-import dagger.android.HasBroadcastReceiverInjector
+import dagger.android.HasAndroidInjector
 import de.metzgore.beansplan.notifications.NotificationHelper
 import de.metzgore.beansplan.util.di.components.AppComponent
 import de.metzgore.beansplan.util.di.components.DaggerAppComponent
@@ -15,13 +13,10 @@ import de.metzgore.beansplan.util.di.modules.ContextModule
 import de.metzgore.beansplan.util.di.modules.ScheduleDaoModule
 import javax.inject.Inject
 
-class BeansPlanApp : Application(), HasActivityInjector, HasBroadcastReceiverInjector {
+class BeansPlanApp : Application(), HasAndroidInjector {
 
     @Inject
-    lateinit var activityDispatchingAndroidInjector: DispatchingAndroidInjector<Activity>
-
-    @Inject
-    lateinit var broadcastDispatchingAndroidInjector: DispatchingAndroidInjector<BroadcastReceiver>
+    lateinit var activityDispatchingAndroidInjector: DispatchingAndroidInjector<Any>
 
     private lateinit var appComponent: AppComponent
 
@@ -32,11 +27,6 @@ class BeansPlanApp : Application(), HasActivityInjector, HasBroadcastReceiverInj
 
         appComponent = createAppComponent()
         appComponent.inject(this)
-
-        if (LeakCanary.isInAnalyzerProcess(this)) {
-            return
-        }
-        LeakCanary.install(this)
     }
 
     private fun createAppComponent(): AppComponent {
@@ -46,11 +36,7 @@ class BeansPlanApp : Application(), HasActivityInjector, HasBroadcastReceiverInj
         }.build()
     }
 
-    override fun activityInjector(): AndroidInjector<Activity>? {
+    override fun androidInjector(): AndroidInjector<Any> {
         return activityDispatchingAndroidInjector
-    }
-
-    override fun broadcastReceiverInjector(): AndroidInjector<BroadcastReceiver> {
-        return broadcastDispatchingAndroidInjector
     }
 }
