@@ -1,6 +1,6 @@
 package de.metzgore.beansplan.data.room
 
-import android.arch.persistence.room.TypeConverter
+import androidx.room.TypeConverter
 import com.google.gson.GsonBuilder
 import de.metzgore.beansplan.data.ShowResponse
 import de.metzgore.beansplan.data.WeeklyScheduleResponse
