@@ -1,6 +1,6 @@
 package de.metzgore.beansplan.util;
 
-import android.arch.lifecycle.LiveData;
+import androidx.lifecycle.LiveData;
 
 public class AbsentLiveData extends LiveData {
     private AbsentLiveData() {
