@@ -1,6 +1,6 @@
 package de.metzgore.beansplan.util.archcomponents
 
-import android.arch.lifecycle.Observer
+import androidx.lifecycle.Observer
 
 class EventObserver<T>(private val onEventUnhandledContent: (T) -> Unit) : Observer<Event<T>> {
     override fun onChanged(event: Event<T>?) {
