@@ -1,6 +1,6 @@
 package de.metzgore.beansplan.util.di.modules
 
-import android.arch.persistence.room.Room
+import androidx.room.Room
 import android.content.Context
 import dagger.Module
 import dagger.Provides
