@@ -1,9 +1,9 @@
 package de.metzgore.beansplan.weeklyschedule
 
-import android.arch.lifecycle.LiveData
-import android.arch.lifecycle.MutableLiveData
-import android.arch.lifecycle.Transformations
-import android.arch.lifecycle.ViewModel
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.Transformations
+import androidx.lifecycle.ViewModel
 import de.metzgore.beansplan.data.Resource
 import de.metzgore.beansplan.data.Status
 import de.metzgore.beansplan.data.room.relations.WeeklyScheduleWithDailySchedules
