@@ -1,8 +1,8 @@
 package de.metzgore.beansplan.weeklyschedule
 
-import android.arch.core.executor.testing.InstantTaskExecutorRule
-import android.arch.lifecycle.MutableLiveData
-import android.arch.lifecycle.Observer
+import androidx.arch.core.executor.testing.InstantTaskExecutorRule
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.Observer
 import de.metzgore.beansplan.LiveDataTestUtil
 import de.metzgore.beansplan.data.Resource
 import de.metzgore.beansplan.data.Status
